all: SSO Phase A — 24 B string header with reserved hash slot
Grows Go's string header from 2 words (16 B) to 3 words (24 B) end-to-
end. Phase A only emits heap-rep strings; the 4-bit length tag in word 2
stays zero and word 1 is reserved for a future cached hash. Inline rep
(tag != 0, up to 15 inline bytes) is left for Phase B. Compiler /
runtime / linker / reflect / vet are all gd-aware.

Design notes pinned at doc/gd/sso-string.md.

Layout (64-bit)
- offset  0: data ptr (heap)            / nil (inline, future)
- offset  8: cached hash, 0 in Phase A  / bytes[0:8] (inline, future)
- offset 16: tag<<60 | len               / tag<<60 | bytes[8:15] packed

Compiler
- internal/abi: StringInlineCap, StringTagShift, StringLenMask.
- compile/internal/types: StringPtrOffset/StringHashOffset/StringLenOffset;
  StringSize = 3*PtrSize; TSTRING intRegs = 3.
- compile/internal/abi: synthString grew to {data, hash, len}.
- compile/internal/ssa: StringMake bumped to 3 args; new StringHash
  selector. dec.rules + generic.rules thread the hash word through
  Load/Store/StringMake/ConstString. expand_calls TSTRING case decomposes
  to 3 slots. decompose.go adds hash-name split. ssagen.storeTypeScalars
  TSTRING now writes hash AND len (this was the long-hunted "missing
  third store" bug — without it heap-stored strings had garbage Len).
- staticdata: TSTRING constants emit (ptr, 0, len) at the three offsets.
- reflectdata/map: assertion uses types.StringSize.
- types2 + go/types Sizes: String → WordSize*3 so vet/asmdecl/Sizeof see
  24 B.
- reflect/abi: String regAssign now uses 3 int regs with mask 0b001.

Runtime
- stringStruct/stringStructDWARF: 3 fields {str, hash uint, len uint};
  internal/runtime/maps/runtime_faststr shadow grown in lockstep.
  Map-elem offsets switch from 2*goarch.PtrSize to typ.ElemOff.
- alg.go strhashFallback: Phase A invariant (tag bits zero) keeps len
  readable as plain uint.
- asm_amd64.s runtime·strhash: read length from 16(AX), not 8(AX).
- internal/abi/map: MapMaxKeyBytes/ElemBytes 128 → 192 so [N≤8]string
  keys still fit inline (otherwise 8*24=192 spilled into the
  per-insertion newobject path).
- print.go / debuglog.go / string.go: stringStruct literals updated.

Linker
- ld/data.go addgostring + addstrdata: emit 3 words (ptr, hash=0, len).
- ld/symtab.go moduledata: empty pluginpath/modulename string slots
  emit 3 zero words instead of 2.

Assembly
- internal/bytealg indexbyte_amd64 / count_amd64 / index_amd64: frame
  sizes ($0-32 → $0-40 etc.) and FP offsets shifted for the 24 B string
  args.
- internal/bytealg compare_amd64 cmpstring: register ABI shuffles
  updated for 3-reg string (a.ptr=AX a.hash=BX a.len=CX,
  b.ptr=DI b.hash=SI b.len=R8).

unsafeheader split
- internal/unsafeheader.String now has two variants behind a build tag:
  string_bootstrap.go (compiler_bootstrap, 2 fields) keeps
  bootstrap-built tools (cmd/compile, vet) safe — they run on host
  16 B strings and must not write through a 3-field view. string_gd.go
  (default, 3 fields) matches the gd runtime layout for reflect and
  embed at user runtime.

vet (asmdecl)
- cmd/vendor/.../asmdecl.go: string component table includes _hash;
  iface table includes the fat-iface inline_real/inline_imag halves.
  rebuild-tools.sh now builds vet from host source via -overlay so the
  shipped vet runs cleanly on host runtime AND knows about gd's 24 B
  string and 32 B fat-iface layouts.

Tooling
- rebuild-tools.sh switches to make.bash -stop=toolchain1: only build
  compile/link/asm/cgo via the bootstrap workspace (host-runtime-linked
  tools), then add a thin bin/go shim that delegates to host go with
  GOROOT pinned to the fork. Avoids ever debugging a fork-compiled-by-
  fork go_bootstrap. vet is rebuilt from host source with an overlay
  for asmdecl + go/types sizes.
- run-tests.sh reworked for the new flow: drives system go +
  GOROOT=fork, filters out fuzz harnesses with -run='^Test', has a
  documented per-package skip list (currently runtime
  TestAtomicAlignment + TestDebugCall*, both tracked separately).

Test status
- HOT set (38 stdlib packages) all green.
- runtime test suite passes with the two skips above (unrelated bugs:
  go/types Importer loops on our export data; debugger-injection asm
  path needs a separate audit).
- reflect/all_test.go TestFuncLayout updated for 3-word strings
  (size 9*PtrSize, stack/gc bitmap [1 0 0 1 0 0 1]).

+++ src/runtime/debuglog.go
@@ -730,7 +730,7 @@ func (r *debugLogReader) printVal() bool {
 		// in this (potentially) nowritebarrier context.
 		str := stringStruct{
 			str: unsafe.Pointer(ptr),
-			len: len,
+			len: uint(len),
 		}
 		s := *(*string)(unsafe.Pointer(&str))
 		print(s)
